refactor: organized regression logic
