+++ src/report-templates/regression/compare-packages.js
